BugID:16959715: reduce codesize for LoRaWAN 4.4.2

+++ network/lorawan_4_4_2/lora/mac/region/RegionAS923.c
@@ -575,7 +575,7 @@ bool RegionAS923AdrNext( AdrNextParams_t* adrNext, int8_t* drOut, int8_t* txPowO
 
 void RegionAS923ComputeRxWindowParameters( int8_t datarate, uint8_t minRxSymbols, uint32_t rxError, RxConfigParams_t *rxConfigParams )
 {
-    double tSymbol = 0.0;
+    DECIMAL tSymbol = 0.0;
     uint32_t radioWakeUpTime;
 
     rxConfigParams->Datarate = datarate;
